Remove duplicate code for query generation
